:recycle: 가게주소 생성 코드 리팩터링
객체를 내부에서 찾지 않고 서비스 레이어에서 가져온 후 매개변수로 전달
